requirements.txt
flask==3.1.0
flask-cors==5.0.0
pymongo==4.10.1
python-dotenv==1.0.0
